backend/requirements_enhanced.txt
# Enhanced requirements with pre-trained models and blockchain
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Blockchain Integration
web3==6.11.3
eth-account==0.9.0
py-solc-x==1.12.0

# AI/ML Models - Pre-trained APIs
google-generativeai==0.3.2
openai==1.3.8
huggingface-hub==0.19.4
transformers==4.36.2
torch==2.1.0
sentence-transformers==2.2.2

# Financial Data APIs
yfinance==0.2.28
alpha-vantage==2.3.1
polygon-api-client==1.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2

# Additional utilities
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
